Enhance universal packaging system & solve compile-time blockers

+++ src/container/runtime.rs
@@ -154,7 +154,9 @@ impl SimpleContainer {
     }
 
     pub fn get_state(&self) -> ContainerState {
-        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) }
+        unsafe {
+            core::mem::transmute(self.state.load(Ordering::SeqCst) as u32)
+        }
     }
 
     pub fn set_state(&self, state: ContainerState) {
@@ -278,6 +280,7 @@ pub trait ContainerRuntime {
 
 /// Runtime statistics
 #[repr(C)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct RuntimeStats {
     pub total_containers: usize,
     pub running_containers: usize,
